sentinel_memory/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
mem0ai>=1.0.0
qdrant-client>=1.8.0
sentence-transformers>=2.6.0
groq>=0.9.0
